requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
werkzeug
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
google-cloud-aiplatform
vertexai
google-cloud-storage
Pillow
